Added edge-case tests for each of the four examples
